@@ -11,14 +11,22 @@
     "react-router": "^3.0.2"
   },
   "devDependencies": {
+    "babel-cli": "^6.23.0",
     "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
+    "babel-loader": "^6.3.2",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.23.0",
+    "html-webpack-plugin": "^2.28.0",
     "http-server": "^0.9.0",
-    "webpack": "^1.14.0"
+    "jquery": "^3.1.1",
+    "node-libs-browser": "^2.0.0",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.4.0",
+    "webpack-merge": "^3.0.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "webpack-dev-server"
   },
   "repository": {
     "type": "git",
